fastapi/pyproject.toml
[tool.poetry]
name = "hybrid-image-search"
version = "0.1.0"
description = "FastAPI backend for hybrid image search using CLIP and pgvector"
authors = ["Admin"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
asyncpg = "^0.29.0"
pgvector = "^0.2.5"
pydantic-settings = "^2.2.1"
pillow = "^10.2.0"
numpy = "^1.26.4"
onnxruntime = "^1.17.1"
python-multipart = "^0.0.9"
tqdm = "^4.66.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
